+++ src/main/java/com/bootcamp/savemypodo/entity/User.java
@@ -1,13 +1,12 @@
 package com.bootcamp.savemypodo.entity;
 
+import com.bootcamp.savemypodo.global.enums.Provider;
+import com.bootcamp.savemypodo.global.enums.Role;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.springframework.security.crypto.password.PasswordEncoder;
-
-import java.time.LocalDateTime;
 
 @Entity
 @Getter
